feat: ScrollContentPresenter occluded padding handling on Skia targets
